requirements.txt
pyrogram==2.0.106
TGCrypto==1.2.5
asyncpg==0.29.0
SQLAlchemy==2.0.30
python-dotenv==1.0.1
openai==1.12.0
python-telegram-bot==20.7
Pillow==10.2.0
tqdm==4.66.1 